Get ready to send and receive more schema
There's a bunch of interesting and somewhat speculative change here.
One of the ideas complicating this code is that, since there might not
be enough memory available for some variable-length allocations, we can
be more resilient by falling back to smaller and smaller lengths until
the allocations succeed.  This may be foolish: the available platforms
and their memory constraints are knowable.