Correct epoch by rounding up to the minute
Event start and end times are transmitted as minutes, then added to the
epoch as multiplied by 60.  If the epoch is not a whole number of
minutes in local time, the result is that in 59 out of 60 cases the
start and end times will be converted to incorrect values.

This fix is not quite accurate as a leap second can throw some start/end
times off by as much as a minute.  The correct solution is to deal with
this in the remote end.